Fix code review findings CR-001 through CR-010
CR-001: createThread helper now takes explicit stackSize parameter
        instead of hardcoding sizeof(g_stackA) for all stacks.
CR-002: _heap_start/_heap_end placed in same linker section to
        guarantee address ordering and prevent negative-size wrap.
CR-003: spiTransferAsync with invalid ID now invokes callback
        (prevents caller deadlock), matching I2C async behavior.
CR-004: IPC Send_FullMailboxAfterResume test documented as host-only
        scenario explaining why mock state invariants are not checked.
CR-005: Added blank line between i2cReadAsync and slave section.
CR-006: Mock validity checks now use < kCount (matching real code)
        instead of <= LastValue to prevent divergence.
CR-007: Added i2cWriteRead invalid ID test.
CR-008: Documented zero-length/invalid-ID async behavior in headers.
CR-009: kI2cCount/kSpiCount changed from uint8_t to uint32_t.
CR-010: Added comment explaining prescaler 99 -> Div256 clamping.

425/425 host tests passing.

+++ hal/src/stm32f4/I2c.cpp
@@ -15,7 +15,7 @@ namespace
     constexpr std::uint32_t kI2c1Base = 0x40005400;
     constexpr std::uint32_t kI2c2Base = 0x40005800;
     constexpr std::uint32_t kI2c3Base = 0x40005C00;
-    constexpr std::uint8_t kI2cCount = 3;
+    constexpr std::uint32_t kI2cCount = 3;
 
     // Register offsets
     constexpr std::uint32_t kCr1 = 0x00;
@@ -747,6 +747,7 @@ namespace hal
         // Generate START
         reg(base + kCr1) |= (1U << kCr1Start);
     }
+
     // --- Slave mode ---
 
     void i2cSlaveInit(I2cId id, std::uint8_t ownAddr,
